# specs/001-mediawatch-saas-platform/contracts/api.openapi.yaml
openapi: 3.0.3
info:
  title: MediaWatch CI API
  description: |
    API REST pour la plateforme MediaWatch CI - Veille médias automatisée pour la Côte d'Ivoire.
    
    Cette API permet aux clients de gérer leurs mots-clés, consulter leurs mentions,
    configurer des alertes, et accéder aux analytics de veille médias.
  version: 1.0.0
  contact:
    name: MediaWatch CI Support
    email: support@mediawatch.ci
servers:
  - url: https://api.mediawatch.ci/v1
    description: Production
  - url: https://staging-api.mediawatch.ci/v1
    description: Staging
  - url: http://localhost:8000/v1
    description: Développement local
...
security:
  - BearerAuth: []
paths:
...
  /keywords:
    get:
      tags: [keywords]
      summary: Lister les mots-clés
      description: Récupère tous les mots-clés de l'organisation de l'utilisateur
      parameters:
        - name: enabled
          in: query
          schema:
            type: boolean
          description: Filtrer par statut activé/désactivé
      responses:
        '200':
          description: Liste des mots-clés
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Keyword'
                  total:
                    type: integer
                  limit_reached:
                    type: boolean
                    description: True si limite de plan atteinte

    post:
      tags: [keywords]
      summary: Créer un mot-clé
      description: Ajoute un nouveau mot-clé à surveiller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  minLength: 1
                  maxLength: 255
                  example: Orange CI
                category:
                  type: string
                  enum: [brand, product, person, competitor, custom]
                  example: brand
                alert_enabled:
                  type: boolean
                  default: true
                alert_threshold:
                  type: number
                  format: float
                  minimum: 0
                  maximum: 1
                  default: 0.3
                  description: Seuil de sentiment négatif pour alertes (0-1)
      responses:
        '201':
          description: Mot-clé créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Keyword'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Mot-clé déjà existant ou limite atteinte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /keywords/{keyword_id}:
    parameters:
      - $ref: '#/components/parameters/KeywordId'
    
    get:
      tags: [keywords]
      summary: Détails d'un mot-clé
      responses:
        '200':
          description: Détails du mot-clé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Keyword'
        '404':
          $ref: '#/components/responses/NotFound'
    
    patch:
      tags: [keywords]
      summary: Modifier un mot-clé
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                category:
                  type: string
                enabled:
                  type: boolean
                alert_enabled:
                  type: boolean
                alert_threshold:
                  type: number
                  format: float
      responses:
        '200':
          description: Mot-clé mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Keyword'
    
    delete:
      tags: [keywords]
      summary: Supprimer un mot-clé
      responses:
        '204':
          description: Mot-clé supprimé

  # ==================== MENTIONS ====================
...
  /exports/csv:
    post:
      tags: [exports]
      summary: Exporter en CSV
      description: Génère un export CSV des mentions filtrées
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: Export CSV (si <500 mentions)
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '202':
          description: Export en cours (si >500 mentions, envoi par email)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Export en cours, vous recevrez un email avec le lien de téléchargement
                  estimated_time:
                    type: integer
                    description: Temps estimé en secondes
  /exports/pdf:
    post:
      tags: [exports]
      summary: Exporter en PDF
      description: Génère un rapport PDF formaté avec graphiques
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '202':
          description: Génération PDF en cours
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  estimated_time:
                    type: integer

  # ==================== ADMIN ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    KeywordId:
      name: keyword_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Keyword:
      type: object
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        category:
          type: string
        enabled:
          type: boolean
        alert_enabled:
          type: boolean
        alert_threshold:
          type: number
          format: float
        total_mentions_count:
          type: integer
        last_mention_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
    ExportRequest:
      type: object
      properties:
        filters:
          type: object
          properties:
            keyword_id:
              type: string
              format: uuid
            source_id:
              type: string
              format: uuid
            sentiment:
              type: string
            date_from:
              type: string
              format: date
            date_to:
              type: string
              format: date
        include_charts:
          type: boolean
          default: false
          description: Inclure graphiques (PDF uniquement)
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'